+++ crates/evm/src/lib.rs
@@ -159,7 +159,7 @@ pub trait ConfigureEvmEnv: Send + Sync + Unpin + Clone + 'static {
 
         // EIP-4844 excess blob gas of this block, introduced in Cancun
         if let Some(excess_blob_gas) = header.excess_blob_gas() {
-            block_env.set_blob_excess_gas_and_price(excess_blob_gas as u64);
+            block_env.set_blob_excess_gas_and_price(excess_blob_gas);
         }
     }
